+++ src/Milochau.Emails/DataAccess/StorageDataAccess.cs
@@ -1,29 +1,45 @@
-﻿using Azure.Storage.Blobs;
+﻿using Azure.Identity;
+using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using Milochau.Emails.Sdk.Models;
+using Milochau.Core.Abstractions;
 using Milochau.Emails.Models.Options;
+using Milochau.Emails.Sdk.Models;
+using System;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace Milochau.Emails.DataAccess
 {
-    public class StorageDataAccess : IStorageDataAccess
+    internal class StorageDataAccess : IStorageDataAccess
     {
-        private readonly StorageOptions options;
+        private readonly CoreHostOptions hostOptions;
+        private readonly ILogger<StorageDataAccess> logger;
+        private readonly string storageAccountUri;
 
-        public StorageDataAccess(IOptions<StorageOptions> options)
+        private const string defaultContainerName = "default";
+
+        public StorageDataAccess(IOptions<CoreHostOptions> hostOptions,
+            EmailsOptions options,
+            ILogger<StorageDataAccess> logger)
         {
-            this.options = options.Value;
+            this.hostOptions = hostOptions.Value;
+            this.logger = logger;
+
+            storageAccountUri = options.StorageAccountUri ?? $"{this.hostOptions.Application.OrganizationName}stg{this.hostOptions.Application.ApplicationName}1{this.hostOptions.Application.HostName}";
         }
 
-        public async Task<Stream> ReadToStreamAsync(StorageAttachment attachment, CancellationToken cancellationToken)
+
+        public async Task<Stream> ReadToStreamAsync(EmailAttachment attachment, CancellationToken cancellationToken)
         {
-            var blobServiceClient = new BlobServiceClient(options.DefaultConnectionString);
-            var blobContainerClient = blobServiceClient.GetBlobContainerClient(attachment.ContainerName);
+            var credential = new DefaultAzureCredential(hostOptions.Credential);
+            var blobServiceClient = new BlobServiceClient(new Uri(storageAccountUri), credential);
+            var blobContainerClient = blobServiceClient.GetBlobContainerClient(defaultContainerName);
             var blobClient = blobContainerClient.GetBlobClient(attachment.FileName);
-            
+
+            logger.LogDebug("Opening a blob from Storage Account...");
             return await blobClient.OpenReadAsync(new BlobOpenReadOptions(allowModifications: false), cancellationToken);
         }
     }