+++ src/Milochau.Emails/Models/Options/StorageOptions.cs
@@ -1,7 +1,0 @@
-﻿namespace Milochau.Emails.Models.Options
-{
-    public class StorageOptions
-    {
-        public string DefaultConnectionString { get; set; }
-    }
-}